+++ changelog.md
@@ -1,1 +1,2 @@
-- Update to 1.21.6
+- Add widget renderer widgets.
+- Add ability to check for other mouse button clicks in buttons.